requirements.txt
Flask
flask_sqlalchemy
flask_socketio
pillow
numpy
opencv-python
matplotlib
psycopg2
datetime
scikit-image
scikit-learn 
pandas
gunicorn
Werkzeug
seaborn
python-dotenv